Fixed incorrect cast of update frequency
Former-commit-id: 0d426d020e8a911be3a54adde1a6e99cc074d8bb

+++ src/config-tool/ConfigTool/src/org/hyperion/hypercon/gui/ColorSmoothingPanel.java
@@ -135,7 +135,7 @@ public class ColorSmoothingPanel extends JPanel {
 		@Override
 		public void stateChanged(ChangeEvent e) {
 			mColorConfig.mSmoothingTime_ms = (Integer)mTimeSpinner.getValue();
-			mColorConfig.mSmoothingUpdateFrequency_Hz = (Integer)mUpdateFrequencySpinner.getValue();
+			mColorConfig.mSmoothingUpdateFrequency_Hz = (Double)mUpdateFrequencySpinner.getValue();
 		}
 	};
 }
